refactor : test 코드 패키지 분리
