+++ src/test/java/lotto/domain/AmountTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
